Fixed Prices's value setter method

+++ catalog/src/main/java/com/netcracker/parfenenko/entities/Price.java
@@ -14,7 +14,7 @@ public class Price {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
